examples: add distributed_join example
This change adds a distributed_join example query which can be run with `cargo run --bin distributed_join_example`. For now (since `ArrowFlightReadExec` is not implemented, it does a non distributed query). The top section in the example can be commented out to make it distributed.

It also adds a small tpch test dataset in testdata/tpch generated using a 0.001 scale factor (largest table is 200kb).