flight_service: add TrailingFlightDataStream
This new type of streams allows you to "chain" streams together. This will be used to "chain" FlightData
containing metrics to the stream corresponding to the last executed partition in a task.

Streams already offer a `chain()` API but it lacks a feature we need - a callback to decide whether
or not to collect and append metrics.

When a stream is complete, as a part of the proposed metrics protocol, the ArrowFlightEndpoint needs to check if
this is the last stream/partition. This synchronization point between N streams for N partitions will
have to be managed in some shared state between the stream and the ArrowFlightEndpoint. In this callback, we
can check if this is the last partition. If so, we can collect metrics and send them in the trailing stream.
For more details, see the draft implementation [here](https://github.com/datafusion-contrib/datafusion-distributed/pull/139/files#diff-fa3e517ceea7f93b2d50873bcdf7f48f6110a5cf8b25a4a8df338f7d71dc6fdb)

One alternative option considered was to have the callback generate `FlightData` which can be sent on
the stream. Ultimately, it seemed cleaner to just return another stream.

Informs: #123