metrics: make MetricsWrapperExec transparent
Before this change, `name()` and `as_any()` on `MetricsWrapperExec` did not delegate calls
to the wrapped node. This was a problem because a user could not traverse the plan and
extract metrics from specific nodes.

After this change, those methods are delegated so a user can now extract metrics
from specific nodes.