add controller package and agent CLI
This change relates to https://github.com/datafusion-contrib/datafusion-distributed/issues/214. It adds a new package to this crate called `datafusion-distributed-controller` which will contain control-plane-related tooling for running a distributed datafusion clusters. This utility will be used for large-scale integration tests and benchmarking. This package does not necessarily need to contain rust code. We can

The proposed architecture is:
- have a local CLI "controller" which can interface with agents running on workers
- the agents can run code on the machine (ex. download data from blob store, start arrow flight server, execute sql on the cluster etc.)

This change adds an agent. The controller will be added in a future commit. Usually an agent runs on a port and is interfaced with via RPC, but since this utility is just for testing, this change proposes that the agent can be interfaced with via a CLI.