fix(ttl_map): two minor fixes for ttl_map configuration parsing
This commit adds two small changes to improve how we parse TTLMapConfigs.

1. It strictly enforces that `tick` is an integer multiple of `ttl`. Previously,
we only recommened that in comments, with the consequence that:

```
true_ttl = desired_ttl - (desired_ttl % tick) <= desired_ttl
```
In the worst case, if `tick` is close to `ttl / 2` then the true `ttl` would
be equivalent to `tick`. For example, if the requested `ttl` was 60 seconds, and the tick
was 31 seconds, we would only allocate a single list of inputs to delete and free them every
tick.

2. There was a small bug in the validation logic that did not correctly validate that tick
was nonzero due to an incorrect application of DeMorgan's law. This commit fixes that to
correctly check that `tick` is nonzero.